+++ src/test/java/org/folio/entlinks/service/consortium/ConsortiumLinksPropagationServiceTest.java
@@ -1,10 +1,14 @@
 package org.folio.entlinks.service.consortium;
 
 import static java.util.Collections.emptyList;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.folio.entlinks.service.consortium.propagation.ConsortiumPropagationService.PropagationType.CREATE;
+import static org.folio.entlinks.service.consortium.propagation.ConsortiumPropagationService.PropagationType.UPDATE;
 import static org.folio.support.base.TestConstants.TENANT_ID;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.times;
@@ -15,7 +19,6 @@ import java.util.List;
 import java.util.UUID;
 import org.folio.entlinks.domain.entity.InstanceAuthorityLink;
 import org.folio.entlinks.exception.FolioIntegrationException;
-import org.folio.entlinks.service.consortium.propagation.ConsortiumAuthorityPropagationService;
 import org.folio.entlinks.service.consortium.propagation.ConsortiumLinksPropagationService;
 import org.folio.entlinks.service.consortium.propagation.model.LinksPropagationData;
 import org.folio.entlinks.service.links.InstanceAuthorityLinkingService;
@@ -23,6 +26,7 @@ import org.folio.spring.service.SystemUserScopedExecutionService;
 import org.folio.spring.testing.type.UnitTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -40,17 +44,23 @@ class ConsortiumLinksPropagationServiceTest {
   void testPropagateUpdate() {
     var instanceId = UUID.randomUUID();
     var link = new InstanceAuthorityLink();
+    link.setId(1L);
     link.setInstanceId(instanceId);
     List<InstanceAuthorityLink> links = List.of(link);
     final var propagationData = new LinksPropagationData(instanceId, links);
+    final var captor = ArgumentCaptor.forClass(List.class);
 
     doMocks();
-    propagationService.propagate(
-        propagationData, ConsortiumAuthorityPropagationService.PropagationType.UPDATE, TENANT_ID);
+    propagationService.propagate(propagationData, UPDATE, TENANT_ID);
 
     verify(tenantsService).getConsortiumTenants(TENANT_ID);
     verify(executionService, times(3)).executeAsyncSystemUserScoped(any(), any());
-    verify(instanceAuthorityLinkingService, times(3)).updateLinks(instanceId, links);
+    verify(instanceAuthorityLinkingService, times(3)).updateLinks(eq(instanceId), captor.capture());
+
+    assertThat(captor.getValue())
+        .usingRecursiveComparison()
+        .ignoringFields("id")
+        .isEqualTo(links);
   }
 
   @Test
@@ -60,8 +70,7 @@ class ConsortiumLinksPropagationServiceTest {
     doMocks();
 
     var exception = assertThrows(IllegalArgumentException.class,
-      () -> propagationService.propagate(propagationData, ConsortiumAuthorityPropagationService.PropagationType.CREATE,
-        TENANT_ID));
+        () -> propagationService.propagate(propagationData, CREATE, TENANT_ID));
 
     assertEquals("Propagation type 'CREATE' is not supported for links.", exception.getMessage());
   }
@@ -72,8 +81,7 @@ class ConsortiumLinksPropagationServiceTest {
 
     final var propagationData = new LinksPropagationData(null, emptyList());
 
-    propagationService.propagate(
-        propagationData, ConsortiumAuthorityPropagationService.PropagationType.UPDATE, TENANT_ID);
+    propagationService.propagate(propagationData, UPDATE, TENANT_ID);
 
     verify(tenantsService, times(1)).getConsortiumTenants(any());
     verify(executionService, times(0)).executeAsyncSystemUserScoped(any(), any());